исправление предыдущих замечаний
