+++ src/components/form-review/const.ts
@@ -1,5 +1,4 @@
 const MIN_LENGTH_COMMENT = 50;
 const MAX_LENGTH_COMMENT = 300;
-const MIN_STARS_COMMENT = 1;
 
-export {MAX_LENGTH_COMMENT, MIN_STARS_COMMENT, MIN_LENGTH_COMMENT};
+export {MAX_LENGTH_COMMENT, MIN_LENGTH_COMMENT};
